Renamed the line menu component file
